refactor: split gateway GPU status helpers
